Changing the message type from 12 bytes to 8 bytes.

+++ one_net/app/one_net_application.c
@@ -217,7 +217,7 @@ BOOL is_broadcast_did(const on_encoded_did_t* did)
     \return FALSE if not parsed successfully
 */
 BOOL on_parse_app_pld(const UInt8* const payload, UInt8* const src_unit,
-  UInt8* const dst_unit, ona_msg_class_t* const msg_class, UInt16* const
+  UInt8* const dst_unit, ona_msg_class_t* const msg_class, UInt8* const
   msg_type, UInt16* const msg_data)
 {
     if(!payload || !src_unit || !dst_unit || !msg_class || !msg_type ||
